Extract obs converters to own module
