--- src/ert/export_observations/__init__.py
@@ -1,3 +1,0 @@
-from .export_observations import export_observations
-
-__all__ = ["export_observations"]
