+++ src/ert/observation_converters/__init__.py
@@ -1,0 +1,4 @@
+from .convert_observations import export_observations
+from .history_to_summary import run_convert_observations
+
+__all__ = ["export_observations", "run_convert_observations"]
