+++ src/ert/observation_converters/summary_to_bulk_config.py
