+++ tests/ert/performance_tests/test_memory_usage.py
@@ -14,7 +14,6 @@ import py
 import pytest
 import xtgeo
 
-from ert.__main__ import run_convert_observations
 from ert.analysis import (
     enif_update,
     smoother_update,
@@ -23,6 +22,7 @@ from ert.config import ErtConfig, ObservationSettings
 from ert.config._create_observation_dataframes import create_observation_dataframes
 from ert.mode_definitions import ENSEMBLE_SMOOTHER_MODE
 from ert.namespace import Namespace
+from ert.observation_converters.history_to_summary import run_convert_observations
 from ert.sample_prior import sample_prior
 from ert.storage import open_storage
 from tests.ert.performance_tests.performance_utils import make_poly_example
